feat(examples): reshape allowlist as BYOI happy-path template (#645)
Add toolbox-hello workload, single entry run.py, and nginx/pi-style README
(build → register → configure → run). Move advanced demos under extras/.
