+++ examples/agent-tool-allowlist-sandbox/extras/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2024 Tencent Inc.
+# SPDX-License-Identifier: Apache-2.0
