+++ examples/agent-tool-allowlist-sandbox/extras/_path.py
@@ -1,0 +1,12 @@
+# Copyright (c) 2024 Tencent Inc.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Put the example root on sys.path (import this first in extras scripts)."""
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+_ROOT = Path(__file__).resolve().parents[1]
+if str(_ROOT) not in sys.path:
+    sys.path.insert(0, str(_ROOT))
