+++ examples/agent-tool-allowlist-sandbox/extras/tool_allowlist_allow.py
@@ -11,6 +11,8 @@ argv allow ≠ network allow.
 
 import os
 
+import _path  # noqa: F401
+
 from e2b_code_interpreter import Sandbox
 
 from env_utils import load_local_dotenv
