+++ examples/agent-tool-allowlist-sandbox/extras/tool_allowlist_checkpoint.py
@@ -14,6 +14,8 @@ Not an LLM agent. Flow:
 
 from __future__ import annotations
 
+import _path  # noqa: F401  # example root on sys.path
+
 import os
 
 from e2b_code_interpreter import Sandbox
