+++ examples/agent-tool-allowlist-sandbox/workspace/README.txt
@@ -1,2 +1,6 @@
-# Agent toolbox workspace (inside the BYOI image).
-# Prefer: cube-tool <name> [args...]  — guest re-checks /etc/cube-sandbox/tool-profile.txt
+# Restricted toolbox workspace (inside the BYOI image).
+#
+# Prefer: cube-tool toolbox-hello
+#   → writes /workspace/out/hello.txt and prints WORKLOAD_OK
+#
+# Guest re-checks /etc/cube-sandbox/tool-profile.txt via cube-tool.
